v3/backend/requirements.txt
flask
python-dotenv
requests
pandas
spotipy
openpyxl